+++ src/KOKKOS/pair_oxdna2_dh_kokkos.cpp
@@ -354,7 +354,7 @@ void PairOxdna2DhKokkos<DeviceType>::operator()(TagPairOxdna2DhCompute<OXDNAFLAG
 
     if (rsq > d_cutsq_dh_c(atype, btype)) continue; // Note the switch of sign, > vs <=, due to using continue
 
-    r = sqrt(rsq);
+    r = sqrtf(rsq);
     rinv = 1.0/r;
 
     if (r <= d_cut_dh_ast(atype, btype)) {
